remove observer pattern from base component
